Broke things apart.
This made no changes other than to take the previous work and break it
into semantically meaningful files/pieces.

Now, each workbook has a file.

The IR work has its own file.
