+++ backend/audit/fixtures/single_audit_checklist.py
@@ -13,7 +13,7 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 
 from faker import Faker
 
-import audit.excel
+import importlib
 import audit.validators
 
 from audit.fixtures.excel import FORM_SECTIONS
@@ -173,10 +173,9 @@ def _post_create_federal_awards(this_sac, this_user):
     )
     excel_file.full_clean()
     excel_file.save()
-
     # TODO: refactor the upload handling from the post view into a
     # function so we can call it here instead of aping it.
-    audit_data = audit.excel.extract_federal_awards(excel_file.file)
+    audit_data = importlib.extract_federal_awards(excel_file.file)
     audit.validators.validate_federal_award_json(audit_data)
     this_sac.federal_awards = audit_data
     this_sac.save()
