+++ backend/audit/test_excel.py
@@ -4,8 +4,9 @@ from django.core.exceptions import ValidationError
 from django.test import SimpleTestCase
 from openpyxl import load_workbook
 from openpyxl.cell import Cell
+ 
 
-from audit.excel import (
+from importlib import (
     ExcelExtractionError,
     extract_federal_awards,
     extract_findings_text,
